% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CivilTwilight.R
\name{CivilTwilight}
\alias{CivilTwilight}
\title{Return civil twilight times from the naval academy website}
\usage{
CivilTwilight(x)
}
\arguments{
\item{URL}{to desired locations civil twilight times}
}
\value{
A dataframe of civil twilight times
}
\description{
Return civil twilight times from the naval academy website
}
\examples{

url <- "https://aa.usno.navy.mil/calculated/rstt/year?ID=AA&year=2024&task=2&lat=39.2805&lon=-76.5940&label=Baltimore\%2C+MD&tz=5.00&tz_sign=-1&submit=Get+Data"
TwilightTable <- CivilTwilight(x=url)
}
