% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FortyEightHour.R
\name{FortyEightHour}
\alias{FortyEightHour}
\title{Returns data from the forty eight hour forecast}
\usage{
FortyEightHour(x)
}
\arguments{
\item{x}{URL to the 48 hour forecast}
}
\value{
A data.frame a data
}
\description{
Returns data from the forty eight hour forecast
}
\examples{

url <- "https://forecast.weather.gov/MapClick.php?lat=39.2736&lon=-76.6264&unit=0&lg=english&FcstType=digital"
HourlyWeatherForecast <- FortyEightHour(url)
}
