% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_Buoys.R
\name{NOAA_Buoys}
\alias{NOAA_Buoys}
\title{Water Temp and Wind Speed from the NOAA stations in the Harbor}
\usage{
NOAA_Buoys(x)
}
\arguments{
\item{x}{An URL for the particular NOAA buoy}
}
\value{
A dataframe of relavent information
}
\description{
Water Temp and Wind Speed from the NOAA stations in the Harbor
}
\examples{

url <- "https://www.ndbc.noaa.gov/data/realtime2/BLTM2.txt"
FtMcHenry <- NOAA_Buoys(x="https://www.ndbc.noaa.gov/data/realtime2/BLTM2.txt")
}
