% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RainfallHistorical.R
\name{RainfallHistorical}
\alias{RainfallHistorical}
\title{Processes historical rainfall data}
\usage{
RainfallHistorical(url)
}
\arguments{
\item{url}{The URL to the weather.gov forecast}
}
\value{
A data.frame of data
}
\description{
Processes historical rainfall data
}
