% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Streptoprocne-package.R
\docType{package}
\name{Streptoprocne-package}
\alias{Streptoprocne}
\alias{Streptoprocne-package}
\title{Streptoprocne: A for-shits-and-giggles side project.}
\description{
To learn more about using the Streptoprocne package, please view our only vignette:
}
\seealso{
Useful links:
\itemize{
  \item \url{https://davidrach.github.io/Streptoprocne/}
  \item Report bugs at \url{https://github.com/DavidRach/Streptoprocne/issues}
}

}
\author{
\strong{Maintainer}: David Rach \email{drach@som.umaryland.edu} (\href{https://orcid.org/0000-0003-2235-5585}{ORCID})

}
\keyword{internal}
