% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SunriseSunset.R
\name{SunriseSunset}
\alias{SunriseSunset}
\title{Return sunrise sunset times from naval academy website}
\usage{
SunriseSunset(x)
}
\arguments{
\item{URL}{to desired locations sunrise sunset table}
}
\value{
A dataframe of sunrise sunset times
}
\description{
Return sunrise sunset times from naval academy website
}
\examples{
NULL
}
