% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SunriseSunset.R
\name{SunriseSunset}
\alias{SunriseSunset}
\title{Return sunrise sunset times from naval academy website}
\usage{
SunriseSunset(x)
}
\arguments{
\item{URL}{to desired locations sunrise sunset table}
}
\value{
A dataframe of sunrise sunset times
}
\description{
Return sunrise sunset times from naval academy website
}
\examples{

url <- "https://aa.usno.navy.mil/calculated/rstt/year?ID=AA&year=2024&task=0&lat=39.2805&lon=-76.5940&label=Baltimore\%2C+MD&tz=5.00&tz_sign=-1&submit=Get+Data"
SunriseTable <- Streptoprocne::SunriseSunset(x=url)
}
